+++ src/main/kotlin/choliver/neapi/ScraperUtils.kt
@@ -1,32 +1,9 @@
 package choliver.neapi
 
-import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 import java.net.URI
 import java.net.URISyntaxException
 import kotlin.math.round
-
-data class ShopifyItemDetails(
-  val title: String,
-  val url: URI,
-  val thumbnailUrl: URI,
-  val price: Double,
-  val available: Boolean
-)
-
-fun Document.shopifyItems() = select(".product-card").map {
-  ShopifyItemDetails(
-    title = it.textFrom(".product-card__title"),
-    url = it.hrefFrom(".grid-view-item__link"),
-    thumbnailUrl = it.srcFrom("noscript .grid-view-item__image")
-      .toString()
-      .replace("@2x", "")
-      .replace("\\?.*".toRegex(), "")
-      .toUri(),
-    price = it.priceFrom(".price-item--sale"),
-    available = "price--sold-out" !in it.selectFirst(".price").classNames()
-  )
-}
 
 fun List<ScrapedItem>.bestPricedItems() = groupBy { it.name }
   .values
@@ -42,6 +19,8 @@ fun Element.attrFrom(cssQuery: String = ":root", attr: String) = selectFrom(cssQ
   .ifBlank { throw ScraperException("Attribute blank or not present: ${attr}") }!!
 fun Element.selectFrom(cssQuery: String) = selectFirst(cssQuery)
   ?: throw ScraperException("Element not present: ${cssQuery}")
+fun Element.selectMultipleFrom(cssQuery: String) = select(cssQuery)
+  .ifEmpty { throw ScraperException("Element(s) not present: ${cssQuery}") }!!
 
 fun String.extract(regex: String) = regex.toRegex().find(this)?.groupValues
 
@@ -52,7 +31,7 @@ fun String.toTitleCase(): String = split(" ").joinToString(" ") {
 // I *know* this doesn't really work for floating-point.  But it's good enough for our purposes.
 fun Double.divideAsPrice(denominator: Int) = round(100 * this / denominator) / 100
 
-private fun String.toUri() = try {
+fun String.toUri() = try {
   URI(this)
 } catch (e: URISyntaxException) {
   throw ScraperException("URL syntax error: ${this}", e)
